+++ scripts/update_version.sh
@@ -4,7 +4,9 @@ set -e
 
 source scripts/utils.sh
 
-git stash --include-untracked
+if [ -n "$(git diff HEAD)" ]; then
+  error "There are uncommitted changes in your workspace"
+fi
 
 $(git config core.editor) scripts/changelog.ts
 
@@ -29,5 +31,3 @@ yarn format
 print_header "Time to commit 🚀️"
 git add -p
 git commit
-
-git stash pop